Fixed and tested the code
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   replacement/lfu/lfu.cc
	modified:   replacement/lip/lip.cc
	modified:   replacement/mru/mru.cc
	modified:   replacement/nmru/nmru.cc
	modified:   replacement/pseudoLRU/pseudoLRU.cc

+++ replacement/lfu/lfu.cc
@@ -55,7 +55,7 @@ void CACHE::initialize_replacement()
         LFU_Policy::policies[this] = std::vector<LFU_Policy::LFU>();
 
         // Allocate the policies, one per cache set
-        for(uint32_t idx = 0; idx < NUM_SET; idx++)
+        for(size_t idx = 0; idx < NUM_SET; idx++)
                 LFU_Policy::policies[this].emplace_back(NUM_WAY);
 }
